pip_requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
array_record==0.6.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
av==14.2.0
bddl==3.6.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
daqp==0.7.1
datasets==4.0.0
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
diffuser==0.0.1
diffusers==0.34.0
dill==0.3.8
dlimp @ git+https://github.com/moojink/dlimp_openvla@040105d256bd28866cc6620621a3d5f7b6b91b46
dm-tree==0.1.8
docker-pycreds==0.4.0
docopt==0.6.2
draccus==0.8.0
easydict==1.13
einops==0.8.0
et_xmlfile==2.0.0
etils==1.11.0
evdev==1.9.2
exceptiongroup==1.2.2
executing==2.2.0
Farama-Notifications==0.0.4
fastjsonschema==2.21.1
ffmpeg==1.4
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.1.24
fonttools==4.55.6
frozenlist==1.7.0
fsspec==2024.12.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.9.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.70.0
gym==0.26.2
gym-notices==0.1.0
gymnasium==0.29.1
h5py==3.12.1
hf-xet==1.1.7
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==8.35.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.5.1
json-numpy==2.1.0
jsonlines==4.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupytext==1.17.2
keras==2.15.0
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mediapy==1.2.2
mergedeep==1.3.4
mink==0.0.11
ml-dtypes==0.2.0
mpmath==1.3.0
mujoco==3.3.4
multidict==6.6.3
multiprocess==0.70.16
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
num2words==0.5.14
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.11.0.86
OpenEXR==3.3.2
openpyxl==3.1.5
opt_einsum==3.4.0
packaging==24.2
pandas==2.3.1
parso==0.8.4
peft==0.13.0
pexpect==4.9.0
pfzy==0.3.4
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.6.0
portalocker==3.2.0
promise==2.3
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.20.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocoevalcap==1.2
pycocotools==2.0.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pynput==1.8.1
PyOpenGL==3.1.9
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-xlib==0.33
pytz==2025.2
PyYAML==6.0.2
pyyaml-include==1.4.1
pyzmq==26.4.0
qpsolvers==4.8.0
qwen-vl-utils==0.0.10
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
robosuite==1.4.0
rpds-py==0.26.0
rsa==4.9
rtree==1.4.0
ruckig==0.15.3
safetensors==0.5.2
sapien==2.2.2
scipy==1.15.1
sentencepiece==0.1.99
sentry-sdk==2.20.0
setproctitle==1.3.4
six==1.17.0
smmap==5.0.2
soupsieve==2.7
stack-data==0.6.3
sty==1.0.6
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-graphics==2021.12.3
tensorflow-io==0.37.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
termcolor==2.5.0
timm==0.9.10
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.0
transforms3d==0.4.2
trimesh==4.6.0
triton==2.2.0
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
validators==0.35.0
wandb==0.19.4
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.14.1
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zipp==3.21.0
